initialize and design Cart section. - added all functionality that need for initialize cart fragment in application. - fixed and improve NikyExceptionMapper.kt.

+++ app/src/main/java/com/abproject/niky/model/dataclass/Cart.kt
@@ -6,9 +6,9 @@ data class Cart(
     @SerializedName("cart_items")
     val cartItems: List<CartItem>,
     @SerializedName("payable_price")
-    val payable_price: Int,
+    val payable_price: Long,
     @SerializedName("shipping_cost")
-    val shippingCost: Int,
+    val shippingCost: Long,
     @SerializedName("total_price")
-    val totalPrice: Int,
+    val totalPrice: Long,
 )
