+++ app/src/main/java/com/abproject/niky/model/dataclass/PurchaseDetail.kt
@@ -1,7 +1,7 @@
 package com.abproject.niky.model.dataclass
 
 data class PurchaseDetail(
-    var totalPrice: Int,
-    var payablePrice: Int,
-    var shippingCost: Int,
+    var totalPrice: Long,
+    var payablePrice: Long,
+    var shippingCost: Long,
 )
